# Faster label loading.