--- a/compiler/rustc_codegen_ssa/Cargo.toml
+++ b/compiler/rustc_codegen_ssa/Cargo.toml
@@ -36,5 +36,12 @@ rustc_target = { path = "../rustc_target" }
 rustc_session = { path = "../rustc_session" }
 
 # Yorick
-ykpack = { git = "https://github.com/softdevteam/yk" }
+bincode = "1.3.1"
+gimli = "0.23.0"
 indexmap = "1.5.2"
+ykpack = { git = "https://github.com/softdevteam/yk" }
+
+[dependencies.object]
+version = "0.22.0"
+default-features = false
+features = ["read_core", "elf"]