--- a/compiler/rustc_codegen_ssa/src/sir.rs
+++ b/compiler/rustc_codegen_ssa/src/sir.rs
@@ -900,7 +900,8 @@ pub mod labels {
         process::Command,
     };
 
-    fn split_symbol(s: &str) -> (String, u32) {
+    /// Splits a Yorick mapping label name into its constituent fields.
+    fn split_label_name(s: &str) -> (String, u32) {
         let data: Vec<&str> = s.split(':').collect();
         debug_assert!(data.len() == 3);
         let sym = data[1].to_owned();
@@ -908,6 +909,7 @@ pub mod labels {
         (sym, bb_idx)
     }
 
+    /// Add a Yorick label section to the specified executable.
     pub fn add_yk_label_section(exe_path: &Path) {
         let labels = extract_dwarf_labels(exe_path).unwrap();
         let mut tempf = tempfile::NamedTempFile::new().unwrap();
@@ -915,6 +917,7 @@ pub mod labels {
         add_section(exe_path, tempf.path());
     }
 
+    /// Copies the bytes in `sec_data_path` into a new Yorick label section of an executable.
     fn add_section(exe_path: &Path, sec_data_path: &Path) {
         let mut out_path = PathBuf::from(exe_path);
         out_path.set_extension("with_labels");
@@ -932,6 +935,8 @@ pub mod labels {
         std::fs::rename(out_path, exe_path).unwrap();
     }
 
+    /// Walks the DWARF tree of the specified executable and extracts Yorick location mapping
+    /// labels. The labels are returned in a vector sorted by the file offset in the executable.
     fn extract_dwarf_labels(exe_filename: &Path) -> Result<Vec<ykpack::SirLabel>, gimli::Error> {
         let file = fs::File::open(exe_filename).unwrap();
         let mmap = unsafe { memmap::Mmap::map(&file).unwrap() };
@@ -941,8 +946,6 @@ pub mod labels {
         } else {
             gimli::RunTimeEndian::Big
         };
-        let mut labels = Vec::new();
-
         let loader = |id: gimli::SectionId| -> Result<&[u8], gimli::Error> {
             Ok(object
                 .section_by_name(id.name())
@@ -959,6 +962,7 @@ pub mod labels {
         let dwarf = dwarf_cow.borrow(&borrow_section);
         let mut iter = dwarf.units();
         let mut subaddr = None;
+        let mut labels = Vec::new();
         while let Some(header) = iter.next()? {
             let unit = dwarf.unit(header)?;
             let mut entries = unit.entries();
@@ -966,17 +970,16 @@ pub mod labels {
                 if entry.tag() == gimli::DW_TAG_subprogram {
                     if let Some(_name) = entry.attr_value(gimli::DW_AT_linkage_name)? {
                         if let Some(lowpc) = entry.attr_value(gimli::DW_AT_low_pc)? {
-                            let addr = match lowpc {
-                                gimli::AttributeValue::Addr(v) => v as u64,
-                                _ => {
-                                    panic!("Error reading dwarf information. Expected type 'Addr'.")
-                                }
-                            };
-                            // We can not accurately insert labels at the beginning of functions,
-                            // because the label is offset by the function headers. We thus simply
-                            // remember the subprogram's address so we can later assign it to the first
-                            // block (ending with '_0') of this subprogram.
-                            subaddr = Some(addr);
+                            if let gimli::AttributeValue::Addr(v) = lowpc {
+                                // We can not accurately insert labels at the beginning of
+                                // functions, because the label is offset by the function headers.
+                                // We thus simply remember the subprogram's address so we can later
+                                // assign it to the first block (ending with '_0') of this
+                                // subprogram.
+                                subaddr = Some(v as u64);
+                            } else {
+                                panic!("Error reading dwarf information. Expected type 'Addr'.")
+                            }
                         }
                     }
                 } else if entry.tag() == gimli::DW_TAG_label {
@@ -985,16 +988,10 @@ pub mod labels {
                             let s = es.to_string()?;
                             if s.starts_with("__YK_") {
                                 if let Some(lowpc) = entry.attr_value(gimli::DW_AT_low_pc)? {
-                                    let addr = match lowpc {
-                                        gimli::AttributeValue::Addr(v) => v as u64,
-                                        _ => panic!(
-                                            "Error reading dwarf information. Expected type 'Addr'."
-                                        ),
-                                    };
                                     if subaddr.is_some() && s.ends_with("_0") {
-                                        // This is the first block of the subprogram. Assign its label
-                                        // to the subprogram's address.
-                                        let (fsym, bb) = split_symbol(s);
+                                        // This is the first block of the subprogram. Assign its
+                                        // label to the subprogram's address.
+                                        let (fsym, bb) = split_label_name(s);
                                         labels.push(ykpack::SirLabel {
                                             off: usize::try_from(subaddr.unwrap()).unwrap(),
                                             symbol_name: fsym,
@@ -1002,12 +999,18 @@ pub mod labels {
                                         });
                                         subaddr = None;
                                     } else {
-                                        let (fsym, bb) = split_symbol(s);
-                                        labels.push(ykpack::SirLabel {
-                                            off: usize::try_from(addr).unwrap(),
-                                            symbol_name: fsym,
-                                            bb,
-                                        });
+                                        let (fsym, bb) = split_label_name(s);
+                                        if let gimli::AttributeValue::Addr(v) = lowpc {
+                                            labels.push(ykpack::SirLabel {
+                                                off: usize::try_from(v as u64).unwrap(),
+                                                symbol_name: fsym,
+                                                bb,
+                                            });
+                                        } else {
+                                            panic!(
+                                                "Error reading dwarf information. Expected type 'Addr'."
+                                            );
+                                        }
                                     }
                                 } else {
                                     // Ignore labels that have no address.