Companion PR ~~coming soon~~ https://github.com/softdevteam/yk/pull/170

A significant cost of starting the JIT is loading our mapping labels from DWARF.

This change extracts the labels (at compile time) into a faster to load format.

Rough benchmarking of a full yk test suite run:

before:
```
	    Mean        Std.Dev.    Min         Median      Max
real        22.323      0.305       21.924      22.303      22.858
user        127.221     0.266       126.793     127.324     127.580
sys         8.258       0.734       7.281       8.002       9.156
```

after:
```
            Mean        Std.Dev.    Min         Median      Max
real        9.082       0.544       8.398       9.157       9.896
user        98.287      0.393       97.876      98.134      98.995
sys         7.054       0.595       6.009       7.202       7.789
```

ci-yk: vext01 faster-labels