@ltratt 2020-12-07
This looks like a nice speed-up! I have a few minor comments, but my only real comment is an overriding one: there's a lack of high-level documentation (e.g. in a function `///` doc comment) explaining what's going on. I think I might struggle to understand what's going on if I look at this in a year's time.