@vext01 2020-12-07
> my only real comment is an overriding one: there's a lack of high-level documentation

It's hairy code that was written some time ago. It's just been moved from ykpack to the compiler.

I can try to comment it up, but I think @ptersilie should review those changes, since he was the original author of this DWARF code.