@vext01 2020-12-07
Comments sorted in 8cc1d23.

I left the `while let` as is.