@ltratt 2020-12-07
I'm not going to die on a hill over this, but I think it's very easy for people to miss "(ascending)".