@vext01 2020-12-07
How about:

> Returns a list of labels in ascending order of file offset.