@ltratt 2020-12-07
I tried that, but I think it's hard to interpret, because it's not necessarily immediately obvious how to relate labels and offsets. That's why I added the "i.e.". Anyway, I'm not going to bikeshed on it further: choose what you think is best and force push a squash ready for merging.