@vext01 2020-12-07
I see what you mean. I can't think of anything better right now.

Squashed.